+++ lib/teamtailor/client.rb
@@ -194,6 +194,20 @@ module Teamtailor
       ).call
     end
 
+    def requisitions(page: 1, include: [])
+      Teamtailor::Request.new(
+        base_url: base_url,
+        api_token: api_token,
+        api_version: api_version,
+        path: '/v1/requisitions',
+        params: {
+          'page[number]' => page,
+          'page[size]' => 30,
+          'include' => include.join(',')
+        }
+      ).call
+    end
+
     private
 
     attr_reader :base_url, :api_token, :api_version