+++ lib/teamtailor/parser.rb
@@ -13,6 +13,7 @@ require 'teamtailor/parser/custom_field'
 require 'teamtailor/parser/custom_field_value'
 require 'teamtailor/parser/referral'
 require 'teamtailor/parser/partner_result'
+require 'teamtailor/parser/requisition'
 
 module Teamtailor
   class Parser
@@ -36,6 +37,7 @@ module Teamtailor
         when 'custom-field-values' then Teamtailor::CustomFieldValue.new(record, included)
         when 'referrals' then Teamtailor::Referral.new(record, included)
         when 'partner-results' then Teamtailor::PartnerResult.new(record, included)
+        when 'requisitions' then Teamtailor::Requisition.new(record, included)
 
         else
           raise Teamtailor::UnknownResponseTypeError, record&.dig('type')